Add admin link to reset http time for a blog
In case the blog incorrectly changes things backwards in time, or
similar situations where the If-Modified-Since header actually points
beyond the end of the blog, we don't get the latest posts. It's really
a problem on the blog side from what I can tell, but this gives the
blog owner a chance to reset the counter and re-fetch the whole blog.