Show blog username/email in listing
For the regular user view, just show which account and email is used for
info.

For the administration view, show the username and email for each
individual blog, to help with tracking down misbehaving blogs.

And finally, in the moderation view, also include the email (it already
had the account name) to make it easier to contact an author.