Remove twitteruser field from blog registration
We can't verify it anymore, and this blocks anybody from saving a blog
that has twitteruser on it.

Twitter does not provide any way to verify the existance of a user
anymore, other than paying $$$$. So we'll just cut the field for now, to
make things work again.

This does not remove *existing* registrations of twitter users, but it
does make it impossible to register new ones.

Issues reported by several different people.